2. MarketAttractiveness/marketattractivenessflow/pyproject.toml
[project]
name = "marketattractivenessflow"
version = "0.1.0"
description = "MarketAttractivenessFlow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.95.0,<1.0.0",
    "agentops>=0.3.23",
    "python-dotenv>=1.0.0"
]

[project.scripts]
kickoff = "marketattractivenessflow.main:kickoff"
plot = "marketattractivenessflow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
